GDB-10512: run the host and legacy workbench on a single server (#1476)
* GDB-10512: Investigate how to run the host and legacy workbench on one server

What
Updates projects to build/start the root-config and legacy-workbench projects on one destination/server.

Why
The root-config is the main project that orchestrates all other ones. The legacy-workbench uses relative paths for endpoints. If both projects run on one server, no changes will be needed for the legacy-workbench to work.

How
Added webpack to the root container of all projects. Added configurations for development and production builds.

# Known issues that will be fixed in next requests
- version replacements for js files
- license report
- when a resource is changed the rebuild is trigered but browser page not refreshed.
