+++ package.json
@@ -5,16 +5,18 @@
     "scripts": {
         "clean": "sh scripts/clean.sh",
         "install": "sh scripts/install.sh",
-        "postinstall": "npm run build:api",
+        "postinstall": "npm run build:api && npm run build:shared",
         "build": "sh scripts/build.sh",
+        "postbuild": "webpack --config=webpack.config.prod.js",
+        "build-dev": "sh scripts/build.sh",
+        "postbuild-dev": "webpack --config=webpack.config.dev.js",
         "build:modules": "concurrently \"npm run build:api\" \"npm run build:shared\"",
         "build:api": "cd packages/api && npm run build",
         "build:shared": "cd packages/shared-components && npm run build",
-        "start": "concurrently \"npm run start-workbench-api\" \"npm run start-legacy-workbench\" \"npm run start-root-config\" \"npm run start-workbench\"",
+        "start": "concurrently \"npm run start-workbench-api\" \"npm run start-workbench\" \"npm run start-root\"",
         "start-workbench-api": "cd packages/api && npm run start",
-        "start-legacy-workbench": "cd packages/legacy-workbench && npm run start",
-        "start-root-config": "cd packages/root-config && npm run start",
-        "start-workbench": "cd packages/workbench && npm run start"
+        "start-workbench": "cd packages/workbench && npm run start",
+        "start-root": "webpack serve --port 9000 --open  --config=webpack.config.dev.js"
     },
     "files": [],
     "repository": {
@@ -36,7 +38,23 @@
     "homepage": "https://github.com/Ontotext-AD/graphdb-workbench#readme",
     "workspaces": [],
     "devDependencies": {
-        "concurrently": "^8.2.2"
+        "@babel/runtime": "^7.19.0",
+        "clean-webpack-plugin": "^4.0.0",
+        "concurrently": "^8.2.2",
+        "copy-webpack-plugin": "^12.0.2",
+        "css-loader": "^7.1.2",
+        "css-minimizer-webpack-plugin": "^7.0.0",
+        "expose-loader": "^5.0.0",
+        "html-loader": "^0.5.5",
+        "ignore-loader": "^0.1.2",
+        "less-loader": "^12.2.0",
+        "mini-css-extract-plugin": "^2.9.0",
+        "style-loader": "^4.0.0",
+        "webpack": "^5.93.0",
+        "webpack-cli": "^5.1.4",
+        "webpack-config-single-spa": "^5.3.1",
+        "webpack-dev-server": "^5.0.4",
+        "webpack-merge-and-include-globally": "^2.3.4"
     },
     "resolutions": {}
 }
