+++ packages/api/src/repository-service.ts
@@ -1,5 +1,5 @@
 export class RepositoryService {
     static getRepositories(): Promise<Response> {
-        return fetch("http://localhost:9001/rest/repositories/all");
+        return fetch("http://localhost:9000/rest/repositories/all");
     }
 }
