+++ packages/legacy-workbench/package.json
@@ -3,9 +3,6 @@
     "version": "2.7.0-RC1",
     "description": "The web application for GraphDB APIs",
     "scripts": {
-        "build": "webpack --config=webpack.config.prod.js",
-        "build-dev": "webpack --config=webpack.config.dev.js",
-        "start": "webpack-dev-server --config=webpack.config.dev.js",
         "lint": "eslint ./src",
         "less:watch": "less-watch-compiler --config=./less-watch-compiler.config.json",
         "less:preprocess": "less-watch-compiler --config=./less-compiler.config.json",
@@ -38,37 +35,15 @@
         "babel-loader": "^8.1.0",
         "babel-polyfill": "^6.26.0",
         "babel-preset-es2015": "^6.24.1",
-        "clean-webpack-plugin": "^4.0.0",
-        "copy-webpack-plugin": "^12.0.2",
         "copyfiles": "^2.4.1",
-        "css-loader": "^3.3.2",
-        "css-minimizer-webpack-plugin": "^7.0.0",
-        "ejs-loader": "^0.5.0",
         "eslint": "^5.16.0",
         "eslint-config-google": "^0.13.0",
-        "expose-loader": "^0.7.5",
-        "extract-loader": "^5.1.0",
-        "file-loader": "^4.3.0",
-        "html-loader": "^0.5.5",
-        "html-webpack-plugin": "^5.6.0",
-        "ignore-loader": "^0.1.2",
         "is-obj": "^2.0.0",
         "less": "^3.10.3",
-        "less-loader": "^5.0.0",
         "license-checker": "^25.0.1",
-        "mini-css-extract-plugin": "^0.8.0",
         "minimist": "^1.2.5",
-        "optimize-css-assets-webpack-plugin": "^5.0.3",
         "puppeteer": "^5.3.1",
-        "sonarqube-scanner": "^3.3.0",
-        "style-loader": "^0.23.1",
-        "url-loader": "^2.3.0",
-        "webpack": "^5.91.0",
-        "webpack-auto-inject-version": "^1.2.2",
-        "webpack-cli": "^5.1.4",
-        "webpack-dev-server": "^5.0.4",
-        "webpack-merge": "^5.10.0",
-        "webpack-merge-and-include-globally": "^2.3.4"
+        "sonarqube-scanner": "^3.3.0"
     },
     "dependencies": {
         "angular-animate": "1.3.8",
