+++ packages/root-config/package.json
@@ -1,13 +1,10 @@
 {
   "name": "@ontotext/root-config",
   "scripts": {
-    "start": "webpack serve --port 9000 --env isLocal --open",
     "lint": "eslint src --ext js",
     "test": "cross-env BABEL_ENV=test jest --passWithNoTests",
     "format": "prettier --write .",
-    "check-format": "prettier --check .",
-    "build": "concurrently npm:build:*",
-    "build:webpack": "webpack --mode=production"
+    "check-format": "prettier --check ."
   },
   "devDependencies": {
     "@babel/core": "^7.23.3",
@@ -27,11 +24,7 @@
     "prettier": "^2.3.2",
     "pretty-quick": "^3.1.1",
     "serve": "^13.0.0",
-    "webpack": "^5.89.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-config-single-spa": "^5.0.0",
-    "webpack-dev-server": "^4.0.0",
-    "webpack-merge": "^5.8.0"
+    "webpack-dev-server": "^4.0.0"
   },
   "dependencies": {
     "@types/jest": "^27.0.1",
