+++ webpack.config.prod.js
@@ -1,12 +1,9 @@
-const PACKAGE = require('./package.json');
 const path = require('path');
 const {merge} = require('webpack-merge');
-const commonConfig = require('./webpack.config.common');
+const commonConfig = require('./webpack.config.common')();
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const {CleanWebpackPlugin} = require('clean-webpack-plugin');
-// const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const CssMinimizerPlugin = require("css-minimizer-webpack-plugin");
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = merge(commonConfig, {
     mode: 'production',
@@ -18,9 +15,11 @@ module.exports = merge(commonConfig, {
         }
     },
     output: {
-        filename: '[name].[contenthash].bundle.js',
+        filename: '[name].[contenthash].js',
         chunkFilename: '[name].[contenthash].bundle.js',
-        path: path.resolve(__dirname, 'dist')
+        path: path.resolve(__dirname, 'dist'),
+        libraryTarget: "system",
+        publicPath: '/'
     },
     module: {
         rules: [
@@ -50,21 +49,7 @@ module.exports = merge(commonConfig, {
         ]
     },
     plugins: [
-        new HtmlWebpackPlugin({
-            template: './src/template.html',
-            favicon: 'src/img/icon.png',
-            templateParameters: {
-                version: PACKAGE.version
-            }
-            // TODO: enable this once completed with the fixes
-            // minify: {
-            //     removeAttributeQuotes: true,
-            //     collapseWhitespace: true,
-            //     removeComments: true
-            // }
-        }),
         new MiniCssExtractPlugin({filename: "[name].[contenthash].css"}),
-        // new OptimizeCssAssetsPlugin(),
         new CssMinimizerPlugin(),
         new CleanWebpackPlugin()
     ]
